+++ changelog.md
@@ -17,6 +17,7 @@ __[maintenance]__
 
 * NLog library is updated from 5.2.7 to 5.2.8.
 * Update certificate information for Node.js installer.
+* Update certificate information for Opera installer.
 
 ## Version 2023.12.11.0